Interim refactor
