+++ app/uk/gov/hmrc/offpayroll/models/PersonalServiceCluster.scala
@@ -16,6 +16,8 @@
 
 package uk.gov.hmrc.offpayroll.models
 
+import uk.gov.hmrc.offpayroll.models.DecisionBuilder.Interview
+
 /**
   * Created by peter on 15/12/2016.
   */
@@ -42,5 +44,16 @@ object PersonalServiceCluster extends Cluster {
 
   override def shouldAskForDecision(clusterAnswers: List[(String, String)]): Boolean = {
     clusterElements.forall((element) => clusterAnswers.exists(a => a._1 == element.questionTag))
+  }
+
+  /**
+    *
+    * Conditional Element based on flow logic of the Cluster
+    *
+    * @param interview
+    * @return
+    */
+  override def getNextElement(interview: Interview): Option[Element] = {
+    clusterElements.headOption
   }
 }
